+++ crates/marine-test-macro-impl/src/marine_test/mod.rs
@@ -20,5 +20,6 @@ mod glue_code_generator;
 mod module_generator;
 mod utils;
 mod modules_linker;
+mod service_generator;
 
 pub use marine_test_impl::marine_test_impl;
